requirements.txt
# Core dependencies for RAG pipeline
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
numpy==1.26.2
pdfplumber==0.10.3
